+++ cloudos/clos.py
@@ -129,7 +129,7 @@ class Cloudos:
         return r
 
     @staticmethod
-    def process_job_list(r, full_data=False):
+    def process_job_list(r, all_fields=False):
         """Process a server response from a self.get_job_list call.
 
         Parameters
@@ -137,7 +137,7 @@ class Cloudos:
         r : requests.models.Response
             The server response. It should contain a field named 'jobs' and
             the required columns (hard-coded in the function).
-        full_data : bool. Default=False
+        all_fields : bool. Default=False
             Whether to return a reduced version of the DataFrame containing
             only the selected columns or the full DataFrame.
 
@@ -176,7 +176,64 @@ class Cloudos:
                    ]
         my_jobs = json.loads(r.content)
         df_full = pd.json_normalize(my_jobs['jobs'])
-        if full_data:
+        if all_fields:
+            df = df_full
+        else:
+            df = df_full.loc[:, COLUMNS]
+        return df
+
+    def get_workflow_list(self, workspace_id):
+        """Get all the workflows from a CloudOS workspace.
+
+        Parameters
+        ----------
+        workspace_id : string
+            The CloudOS workspace id from to collect the workflows.
+
+        Returns
+        -------
+        r : requests.models.Response
+            The server response
+        """
+        data = {"apikey": self.apikey}
+        r = requests.get("{}/api/v1/workflows?teamId={}".format(self.cloudos_url,
+                                                                workspace_id),
+                         params=data)
+        if r.status_code >= 400:
+            raise BadRequestException()
+        return r
+
+    @staticmethod
+    def process_workflow_list(r, all_fields=False):
+        """Process a server response from a self.get_workflow_list call.
+
+        Parameters
+        ----------
+        r : requests.models.Response
+            The server response. It should contain a field named 'workflows' and
+            the required columns (hard-coded in the function).
+        all_fields : bool. Default=False
+            Whether to return a reduced version of the DataFrame containing
+            only the selected columns or the full DataFrame.
+
+        Returns
+        -------
+        df : pandas.DataFrame
+            A DataFrame with the requested columns from the workflows.
+        """
+        COLUMNS = ['_id',
+                   'name',
+                   'archived.status',
+                   'mainFile',
+                   'workflowType',
+                   'repository.name',
+                   'repository.platform',
+                   'repository.url',
+                   'repository.isPrivate'
+                   ]
+        my_workflows = json.loads(r.content)
+        df_full = pd.json_normalize(my_workflows)
+        if all_fields:
             df = df_full
         else:
             df = df_full.loc[:, COLUMNS]