+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 - Adds WDL pipeline support, iteration 2: WDL workflows can be run
 using the regular `cloudos job run` using the new arguments:
     * `--wdl-mainfile`
+    * `--wdl-importsfile`
     * `--cromwell-token`
+- Adds the new argument `--repository-platform` to specify the
+repository platform ('github' or 'bitcucket')
 
 ### 0.1.1 - 2022-07-12
 - Adds WDL pipeline support, iteration 1: cromwell server managing.